requirements.txt
asn1crypto==0.24.0
base58==1.0.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3.1
ecdsa==0.13
Flask==1.0.2
Flask-JSGlue==0.3.1
gitdb2==2.0.4
GitPython==2.1.11
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
ontology-dapp-box==0.0.1
ontology-python-sdk==0.1.3
pycparser==2.18
pycryptodomex==3.6.6
requests==2.19.1
selenium==3.14.0
six==1.11.0
smmap2==2.0.4
urllib3==1.23
Werkzeug==0.14.1
